refactor: move Result and NestedResult into `Result` subnamespace
In doing so, discovered that the `Rule::validate()` method needed to tighten up the `$context` argument.
