+++ src/Rule/CallbackRule.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace Phly\RuleValidation\Rule;
 
-use Phly\RuleValidation\Result;
 use Phly\RuleValidation\Rule;
+use Phly\RuleValidation\ValidationResult;
 
 final class CallbackRule implements Rule
 {
-    /** @var callable(mixed, array, non-empty-string): Result */
+    /** @var callable(mixed, array<non-empty-string, mixed>, non-empty-string): ValidationResult */
     private $callback;
 
-    /** @param callable(mixed, array, non-empty-string): Result $callback */
+    /** @param callable(mixed, array<non-empty-string, mixed>, non-empty-string): ValidationResult $callback */
     public function __construct(
         /** @var non-empty-string */
         private string $key,
@@ -33,8 +33,12 @@ final class CallbackRule implements Rule
         return $this->key;
     }
 
-    /** Validate the value */
-    public function validate(mixed $value, array $context): Result
+    /**
+     * Validate the value
+     *
+     * @param array<non-empty-string, mixed> $context
+     */
+    public function validate(mixed $value, array $context): ValidationResult
     {
         return ($this->callback)($value, $context, $this->key);
     }
