+++ test/RuleSet/RuleSetTest.php
@@ -7,8 +7,8 @@ namespace PhlyTest\RuleValidation\RuleSet;
 use Phly\RuleValidation\Exception\DuplicateRuleKeyException;
 use Phly\RuleValidation\Exception\RequiredRuleWithNoDefaultValueException;
 use Phly\RuleValidation\Exception\ResultSetFrozenException;
-use Phly\RuleValidation\Result;
 use Phly\RuleValidation\Result\CreateMissingValueResult;
+use Phly\RuleValidation\Result\Result;
 use Phly\RuleValidation\ResultSet;
 use Phly\RuleValidation\Rule;
 use Phly\RuleValidation\RuleSet\RuleSet;
