+++ test/TestAsset/CustomResultSet.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhlyTest\RuleValidation\TestAsset;
 
-use Phly\RuleValidation\Result;
+use Phly\RuleValidation\Result\Result;
 use Phly\RuleValidation\ResultSet;
 
 /**
